requirements.txt

# requirements.txt
# Web Framework & Server
fastapi==0.128.0
uvicorn[standard]==0.40.0

# Data Validation & Settings
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

# Security & Auth
PyJWT[cryptography]==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.21

# Database
SQLAlchemy==2.0.45
pymysql==1.1.2
cryptography==46.0.3
alembic==1.18.1
DBUtils==3.1.2

# Utilities
filelock==3.20.3